Extracting the list of expenses to separate component
