+++ src/components/Expenses/ExpenseItem.js
@@ -6,6 +6,7 @@ import './ExpenseItem.css';
 
 const ExpenseItem = (props) => {
     return (
+      <li>
         <Card className="expense-item">
             <ExpenseDate date={props.date} />
             <div className='expense-item__description'>
@@ -13,6 +14,7 @@ const ExpenseItem = (props) => {
                 <div className='expense-item__price'>${props.amount}</div>
             </div>
         </Card>
+      </li>
     );
 }
