+++ src/components/Expenses/ExpensesList.css
@@ -1,0 +1,9 @@
+.expenses-list {
+  list-style: none;
+  padding: 0;
+}
+
+.expenses-list__fallback {
+  color: white;
+  text-align: center;
+}
